Switch key store 'keys' from DYNARRAY to list.
This helps resolve the issue of pointers to keys in a store becoming
invalid in various cases.

+++ include/rnp/rnp.h
@@ -78,15 +78,15 @@ int         rnp_list_packets(rnp_t *, char *, int);
 int rnp_set_key_store_format(rnp_t *, const char *);
 
 /* key management */
-bool  rnp_list_keys(rnp_t *, const int);
-bool  rnp_list_keys_json(rnp_t *, char **, const int);
-bool  rnp_find_key(rnp_t *, const char *);
-char *rnp_get_key(rnp_t *, const char *, const char *);
-char *rnp_export_key(rnp_t *, const char *);
-int   rnp_import_key(rnp_t *, char *);
-bool  rnp_generate_key(rnp_t *);
-int   rnp_secret_count(rnp_t *);
-int   rnp_public_count(rnp_t *);
+bool   rnp_list_keys(rnp_t *, const int);
+bool   rnp_list_keys_json(rnp_t *, char **, const int);
+bool   rnp_find_key(rnp_t *, const char *);
+char * rnp_get_key(rnp_t *, const char *, const char *);
+char * rnp_export_key(rnp_t *, const char *);
+int    rnp_import_key(rnp_t *, char *);
+bool   rnp_generate_key(rnp_t *);
+size_t rnp_secret_count(rnp_t *);
+size_t rnp_public_count(rnp_t *);
 
 /* file management */
 rnp_result_t rnp_process_file(rnp_ctx_t *, const char *, const char *);
